Level 3 Complete

+++ Mak Attack/Assets/Scripts2/RayoScriptLevel2.cs
@@ -37,7 +37,7 @@ public class RayoScriptLevel2 : MonoBehaviour {
     void FixedUpdate()
     {
         PlayerPrefs.SetFloat("health", rayoLife);
-        PlayerPrefs.SetFloat("maxhealth", rayoMaxLife);
+        PlayerPrefs.SetInt("maxhealth", rayoMaxLife);
         PlayerPrefs.SetInt("score", rayoScore);
         PlayerPrefs.SetFloat("damage", rayoDamage);
         PlayerPrefs.SetInt("potions", potionCount);
